refactor: move files into ecs project
Architecture of the project has been updated, and now main features related to the engine itself can be found at [ECS](https://github.com/andywiecko/ECS) project. There are a few more things to refactor with respect to given changes, but it will be performed in the future, this includes renaming, fixing the namespaces fixing the editor classes, etc.